+++ packages/cli/src/util/telemetry/commands/teams/index.ts
@@ -42,10 +42,21 @@ export class TeamsTelemetryClient
     }
   }
 
-  trackCliSubcommandRequest(actual: string) {
-    this.trackCliSubcommand({
-      subcommand: 'request',
-      value: actual,
-    });
+  trackCliSubcommandRequest(actual?: string) {
+    if (actual) {
+      this.trackCliSubcommand({
+        subcommand: 'request',
+        value: actual,
+      });
+    }
+  }
+
+  trackCliSubcommandMembers(actual?: string) {
+    if (actual) {
+      this.trackCliSubcommand({
+        subcommand: 'members',
+        value: actual,
+      });
+    }
   }
 }